// src/common/models/Enums.ts
export enum WeekDay {
    Monday = 'Monday',
    Tuesday = 'Tuesday',
    Wednesday = 'Wednesday',
    Thursday = 'Thursday',
    Friday = 'Friday',
    Saturday = 'Saturday',
    Sunday = 'Sunday'
}

export enum Month {
    January = 'January',
    February = 'February',
    March = 'March',
    April = 'April',
    May = 'May',
    June = 'June',
    July = 'July',
    August = 'August',
    September = 'September',
    October = 'October',
    November = 'November',
    December = 'December',
}

export enum Vertical {
    AllVerticals = "All Verticals",
    SpaceNW = "Space & NW",
    LandFixed = "Land Fixed",
    LandMobility = "Land Mobility",
    Aviation = "Aviation",
    Maritime = "Maritime",
    SNP = "SNP"
}

export enum MarketReadinessResponsibleParty {
    Eutelsat = "Eutelsat",
    DP = "Distribution Partner",
    TP = "Teleport Partner",
    Overall = "Overall"
}

export enum MARequirementResponsibleParty {
    Eutelsat = "Eutelsat",
    DP = "Distribution Partner",
    TP = "Teleport Partner",
}

export enum MARequirementType {
    Generic = "Generic",
    Legal = "Legal",
    NOL = "NOL",
    Credential = "Credential"
}

export enum AuthorityType {
    Organization = "Organization",
    Regulator = 'Regulator',
    Administration = "Administration",
}

export enum LicenceRenewalTerm {
    Fixed = "Fixed Term",
    Permanent = 'Permanent',
    Recurrent = 'Recurrent'
}

export enum SNPType {
    Stargate = "Stargate",
    EutelsatOwned = "Eutelsat Owned",
    Other = "Other"
}

export enum FeeType {
    SNP = "SNP",
    MATypeRequirement = "MA Requirement",
}

export enum FeeChargedBy {
    Organization = "Organization",
    Regulator = "Regulator",
    Administration = "Administration",
    TeleportPartner = "Teleport Partner",
    LawFirm = "Law Firm",
    LegalRepresentative = "Legal Representative"
}

export enum ContactType {
    Eutelsat = "Eutelsat",
    Organization = "Organization",
    Regulator = 'Regulator',
    Administration = "Administration",
    TP = 'Teleport Partner',
    DP = 'Distribution Partner',
    LegalRepresentative = 'Legal Representative',
    LawFirm = 'Law Firm'
}

export enum LegalSvcProviderType {
    LawFirm = "Law Firm",
    LegalRepresentative = 'Legal Representative'
}

export enum RelatedItemType {
    SNP = "SNP",
    TeleportPartner = "Teleport Partner",
    DistributionPartner = "Distribution Partner",
    LegalRepresentative = "Legal Representative",
    LawFirm = "Law Firm",
    Administration = "Administration",
    Regulator = "Regulator",
    Organization = "Organization",
}


export enum UserRole {
    FinanceContributor = 'Finance Contributor',
    FinanceVisitor = 'Finance Visitor',
    Contributor = 'Contributor',
    Owner = 'Owner',
    Visitor = 'Visitor',
    Admin = 'Admin'
}